Restructure resource file locations to Django expected locations (#68)
