+++ csunplugged/config/settings.py
@@ -65,8 +65,7 @@ TEMPLATES = [
     {
         'BACKEND': 'django.template.backends.django.DjangoTemplates',
         'DIRS': [
-            os.path.join(SETTINGS_PATH, 'templates'),
-            os.path.join(SETTINGS_PATH, 'resources/content/')
+            os.path.join(SETTINGS_PATH, 'templates')
         ],
         'APP_DIRS': True,
         'OPTIONS': {
