+++ csunplugged/resources/urls.py
@@ -1,6 +1,6 @@
 from django.conf.urls import url
 
-from . import views
+from .views import views
 
 app_name = 'resources'
 urlpatterns = [
