+++ csunplugged/resources/views/generate_resource_pdf.py
